fix merge error
