+++ src/windowmanagement/WindowLayout/types.ts
@@ -1,0 +1,6 @@
+export type FloatWindowPosition = {
+  offsetX: number;
+  offsetY: number;
+  width: number;
+  height: number;
+};
